+++ choco/src/main/java/btrplace/solver/choco/actionModel/ForgeVMModel.java
@@ -26,15 +26,22 @@ import btrplace.solver.choco.Slice;
 import btrplace.solver.choco.SliceBuilder;
 import btrplace.solver.choco.VMActionModel;
 import choco.cp.solver.CPSolver;
-import choco.cp.solver.variables.integer.IntDomainVarAddCste;
 import choco.kernel.solver.variables.integer.IntDomainVar;
 
 import java.util.UUID;
 
 /**
- * Model an action that forge a VM to put it into the ready state.
- * The VM must has an attribute {@code template} which indicate the template
- * to use to build the VM.
+ * Model an action that forge a VM to put it into the ready state. *
+ * <p/>
+ * The VM must have an attribute (provided by {@link btrplace.model.Model#getAttributes()}
+ * {@code template} that indicate the template identifier to use to build the VM image.
+ * <p/>
+ * The model must provide an estimation of the action duration through a
+ * {@link btrplace.solver.choco.DurationEvaluator} accessible from
+ * {@link btrplace.solver.choco.ReconfigurationProblem#getDurationEvaluators()} with the key {@code ForgeVM.class}
+ * <p/>
+ * If the reconfiguration problem has a solution, a {@link btrplace.plan.event.ForgeVM} action
+ * will inserted into the resulting reconfiguration plan.
  *
  * @author Fabien Hermenier
  */
@@ -46,9 +53,9 @@ public class ForgeVMModel implements VMActionModel {
 
     private IntDomainVar state;
 
-    private IntDomainVar end;
-
     private Slice dSlice;
+
+    private String template;
 
     /**
      * Make a new model.
@@ -58,26 +65,34 @@ public class ForgeVMModel implements VMActionModel {
      * @throws SolverException if an error occurred
      */
     public ForgeVMModel(ReconfigurationProblem rp, UUID e) throws SolverException {
-        /*
-         * We don't make any "real" dslice cause it may impacts the TaskScheduler
-         */
         int d = rp.getDurationEvaluators().evaluate(ForgeVM.class, e);
+        template = rp.getSourceModel().getAttributes().getString(e, "template");
+        if (template == null) {
+            throw new SolverException(rp.getSourceModel(), "Unable to forge the VM '" + e + "'. The required attribute 'template' is missing from the model");
+        }
         CPSolver s = rp.getSolver();
         duration = s.makeConstantIntVar(d);
         state = s.makeConstantIntVar(0);
         vm = e;
 
+        /*
+         * We don't make any "real" dslice cause it may impacts the TaskScheduler
+         * so the hosting variable is set to -1 to be sure the VM is not hosted on a node
+         */
+
         dSlice = new SliceBuilder(rp, e, rp.makeVarLabel("forge(" + e + ").dSlice"))
-                .setEnd(rp.getEnd())
+                .setDuration(duration)
                 .setStart(rp.makeDuration("forge(" + e + ").start"))
+                .setHoster(-1)
                 .build();
         s.post(s.leq(d, dSlice.getDuration()));
-        end = new IntDomainVarAddCste(s, rp.makeVarLabel("forge(" + e + ").start"), dSlice.getStart(), -d);
+        s.post(s.leq(dSlice.getEnd(), rp.getEnd()));
     }
 
     @Override
     public boolean insertActions(ReconfigurationPlan plan) {
-        return true;
+        ForgeVM a = new ForgeVM(vm, getStart().getVal(), getEnd().getVal());
+        return plan.add(a);
     }
 
     @Override
@@ -92,7 +107,7 @@ public class ForgeVMModel implements VMActionModel {
 
     @Override
     public IntDomainVar getEnd() {
-        return end;
+        return dSlice.getEnd();
     }
 
     @Override
@@ -120,4 +135,12 @@ public class ForgeVMModel implements VMActionModel {
         v.visit(this);
     }
 
+    /**
+     * Get the template to use to build the VM.
+     *
+     * @return the template identifier
+     */
+    public String getTemplate() {
+        return template;
+    }
 }